+++ app/Http/Controllers/AdminPanel/CategoryController.php
@@ -132,7 +132,9 @@ class CategoryController extends Controller
     public function destroy(Category $category,$id)
     {
         $data=Category::find($id);
+        if($data->image && Storage::disk('public')->exists($data->image)){
             Storage::delete($data->image);
+        }
         $data->delete();
         return redirect('admin/category');
     }
